+++ src/AdminServiceProvider.php
@@ -123,6 +123,7 @@ final class AdminServiceProvider extends ServiceProvider
 
         $this->loadViewsFrom(__DIR__.'/../resources/views', 'admin');
         $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'admin');
 
         $this->registerAdminGuard();
         $this->registerCommands();